# api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CryptoTrend api",
        "description": "REST API приложения тренды криптовалют",
        "contact": {
            "email": "vahagn99ghukasyan@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Local server"
        }
    ],
    "paths": {
        "/api/v1/auth/otp": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Отправка OTP-кода на телефон",
                "description": "Отправляет одноразовый код подтверждения (OTP) на указанный номер телефона. Для локального окружения возвращает токен из переменных окружения: `AUTH_OTP_NONE_PRODUCTION_ENV_CODE` по умолчанию `123456`",
                "operationId": "sendOtpCode",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "phone"
                                ],
                                "properties": {
                                    "phone": {
                                        "description": "Телефон в формате от 10 до 15 цифр. Для локального окружения используется по умолчанию `79991234567`",
                                        "type": "string",
                                        "example": "79991234567"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Код успешно отправлен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Код отправлен"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации телефона",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Неверный формат телефона"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "Превышен лимит запросов",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Слишком много запросов, попробуйте позже"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Произошла внутренняя ошибка сервера"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/auth/otp/verify": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Подтверждение OTP-кода",
                "description": "Проверяет OTP-код и возвращает токен авторизации при успехе.",
                "operationId": "verifyOtpCode",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "phone",
                                    "code"
                                ],
                                "properties": {
                                    "phone": {
                                        "description": "Телефон",
                                        "type": "string",
                                        "example": "79991234567"
                                    },
                                    "code": {
                                        "description": "OTP-код",
                                        "type": "string",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Успешная верификация",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Успешная верификация"
                                        },
                                        "access_token": {
                                            "type": "string",
                                            "example": "eyJhbGciOi..."
                                        },
                                        "token_type": {
                                            "type": "string",
                                            "example": "Bearer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации или неверный код",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Неверный код"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Произошла ошибка сервера, дождитесь пока менеджер свяжется с вами!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/prices": {
            "get": {
                "tags": [
                    "Crypto"
                ],
                "summary": "Получить последние обновления цен по монетам",
                "description": "Возвращает список монет с последними снимками цены начиная с указанной даты",
                "operationId": "getCryptoLastUpdates",
                "parameters": [
                    {
                        "name": "from",
                        "in": "query",
                        "description": "Дата начала выборки (формат: Y-m-d)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2025-07-01"
                        }
                    },
                    {
                        "name": "vs_currency",
                        "in": "query",
                        "description": "Валюта отображения цен",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "usd",
                            "enum": [
                                "usd"
                            ]
                        }
                    },
                    {
                        "name": "provider",
                        "in": "query",
                        "description": "Провайдер данных",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "coingecko",
                            "enum": [
                                "coingecko"
                            ]
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Количество элементов на странице",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 50
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Номер страницы",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ со списком монет",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Coin"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "last_page": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 50
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 200
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Ошибка валидации"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "Превышен лимит запросов",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Слишком много запросов, попробуйте позже"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Критическая ошибка сервера",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Произошла ошибка сервера, дождитесь пока менеджер свяжется с вами!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Coin": {
                "title": "Coin",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 2
                    },
                    "name": {
                        "type": "string",
                        "example": "Bitcoin"
                    },
                    "symbol": {
                        "type": "string",
                        "example": "BTC"
                    }
                },
                "type": "object"
            },
            "CoinSnapshot": {
                "title": "CoinSnapshot",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 2
                    },
                    "coin_id": {
                        "type": "integer",
                        "example": "2"
                    },
                    "price": {
                        "type": "integer",
                        "example": "15000"
                    },
                    "volume": {
                        "type": "integer",
                        "example": "15000"
                    },
                    "market_cap": {
                        "type": "integer",
                        "example": "15000"
                    },
                    "percent_change_24h": {
                        "type": "integer",
                        "example": "15000"
                    },
                    "fetched_at": {
                        "type": "string",
                        "example": "2025-07-13 00:00:00"
                    },
                    "used_provider": {
                        "type": "string",
                        "example": "coingecko"
                    },
                    "vs_currency": {
                        "type": "string",
                        "example": "usd"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Auth",
            "description": "Auth"
        },
        {
            "name": "Crypto",
            "description": "Crypto"
        }
    ]
}